open-campus-demo-v1/services/conversation_agent/requirements.txt
pydantic
python-dotenv
certifi
